actualizar todo

+++ Parcial1_2B/6_Ejercicios_repaso_1al5/ejercicio4.py
@@ -1,5 +1,5 @@
-num1=int(input("Coloca el primer numero: "))
-num2=int(input("Coloca el segundo numero: "))
+num1=float(input("Coloca el primer numero: "))
+num2=float(input("Coloca el segundo numero: "))
 
 suma= num1 +num2
 resta= num1 - num2
